refactor(cmd): moving to kamaji module
