+++ cmd/kamaji/utils/check_flags.go
